+++ src/parser.js
@@ -330,8 +330,6 @@ function resolveLineIndexes(node, lineIndexes) {
     if (nodeLoc.endLine < 0) {
       nodeLoc.endLine = lineIndexes.length - 1;
     }
-    // Edge case: Trigger Declaration Unit.
-    // Somehow jorje thinks this node ends after the word `trigger`.
   }
   Object.keys(node).forEach(key => {
     if (typeof node[key] === "object") {